pyproject.toml
[tool.poetry]
name = "gtfsdb-seg"
version = "0.0.1"
description = "generate segment tables for stops, routes, route-stops"
authors = ["Frank Purcell <fxpurcell@gmail.com>"]
license = "Mozella 2.x"
readme = "README.md"
packages = [{include = "gtfsdb_seg"}]

[tool.poetry.dependencies]
python = "^3.6"
psycopg2-binary = "*"
# "gtfsdb" = {git = "https://github.com/OpenTransitTools/gtfsdb.git", branch = "rtp"}
"gtfsdb" = { path = "../gtfsdb/", develop = true }
"ott.utils" = { path = "../utils/", develop = true }

[tool.poetry.scripts]
segment = "gtfsdb_seg.loader:main"
segment_stops = "gtfsdb_seg.loader:segment_stops"
segment_routes = "gtfsdb_seg.loader:segment_routes"
